// apps/pharmacy/src/types/staff/doctor.ts
export interface DoctorDTO {
  id: number;
  name: string;
  surname: string | null;
  dob: string;
  employeeId: string | null;
  title: string | null;
  phone: string | null;
  email: string | null;
  notes: string | null;
  department: Department | null;
  designation: StaffDesignation | null;
  createdAt: Date | null;
  createdBy: number | null;
  updatedBy: number | null;
  updatedAt: Date | null;
}

export interface CreateOrUpdateDoctor {
  id?: number;
  name: string;
  surname: string;
  employeeId: string;
  dob: string;
  title?: string | null;
  phone?: string | null;
  email: string;
  notes?: string | null;
  siteId?: number | null;
  designationId?: number | null;
  departmentId?: number | null;
  locationId?: number | null;
}

export interface ExcelDoctorRow {
  Name: string;
  Surname: string;
  "employee ID": string;
  DOB: string;
  Title?: string | null;
  Phone?: number | null;
  Email: string;
  Notes?: string | null;
  Site?: number | null;
  Department?: number | null;
  Location?: number | null;
  Designation?: number | null;
}
// ...
export type StaffEntity = Prisma.StaffGetPayload<{
  include: {
    employee: true;
  };
}>;